Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int x,y,z;
+    float x,y,z;
     float avg;
-    scanf("%d %d %d",&x,&y,&z);
+    scanf("%f %f %f",&x,&y,&z);
     avg=(x+y+z)/3;
     printf("Average: %.2f",avg);
